requirements.txt
gradio
Pillow
numpy
datasets
transformers
pyyaml==5.1
torch==1.8.0+cu101 
torchvision==0.9.0+cu101
detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu101/torch1.8/index.html
pytesseract